requirements-dev.txt
-r requirements.txt
pytest
pytest-django
pytest-xdist  # For parallel test execution
pytest-benchmark  # For performance testing
factory-boy
Faker
coverage
black
isort
flake8
djangorestframework